AI-Contract-Analysis-System/requirements.txt
flask
streamlit
spacy
nltk
scikit-learn
firebase-admin
pdfplumber
python-docx
pandas
numpy
fpdf
